Fast

+++ hittable.h
@@ -11,7 +11,7 @@ struct hit_record {
     point3 start_point;
     vec3 normal;
     shared_ptr<material> mat_ptr;
-    double t;
+    float t;
     bool front_face;
 
     inline void set_face_normal(const ray& r, const vec3& outward_normal) {
@@ -22,7 +22,7 @@ struct hit_record {
 
 class hittable {
     public:
-        virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const = 0;
+        virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
 };
 
 #endif
